global coords is now top down, not recursive up

+++ src/PlugTypes.h
@@ -36,7 +36,7 @@
 
     struct UIPoint {//position
         UIPoint() : x(0), y(0) {}
-        UIPoint(float X, float Y) : x(X), y(Y) {}
+        UIPoint(int X, int Y) : x(X), y(Y) {}
         int x;
         int y;
     };
